+++ packages/dataviews/src/components/dataviews-filters/search-widget.tsx
@@ -13,7 +13,8 @@ import { useState, useMemo, useDeferredValue } from '@wordpress/element';
 import {
 	VisuallyHidden,
 	Icon,
-	privateApis as componentsPrivateApis,
+	Composite,
+	useCompositeStore,
 } from '@wordpress/components';
 import { search, check } from '@wordpress/icons';
 import { SVG, Circle } from '@wordpress/primitives';
@@ -21,14 +22,7 @@ import { SVG, Circle } from '@wordpress/primitives';
 /**
  * Internal dependencies
  */
-import { unlock } from '../../lock-unlock';
 import type { Filter, NormalizedFilter, View } from '../../types';
-
-const {
-	CompositeV2: Composite,
-	CompositeItemV2: CompositeItem,
-	useCompositeStoreV2: useCompositeStore,
-} = unlock( componentsPrivateApis );
 
 interface SearchWidgetProps {
 	view: View;
@@ -108,18 +102,20 @@ function ListBox( { view, filter, onChangeView }: SearchWidgetProps ) {
 				filter.name
 			) }
 			onFocusVisible={ () => {
+				// `onFocusVisible` needs the `Composite` component to be focusable,
+				// which is implicitly achieved via the `virtualFocus: true` option
+				// in the `useCompositeStore` hook.
 				if ( ! compositeStore.getState().activeId ) {
 					compositeStore.move( compositeStore.first() );
 				}
 			} }
-			render={ <Ariakit.CompositeTypeahead store={ compositeStore } /> }
+			render={ <Composite.Typeahead /> }
 		>
 			{ filter.elements.map( ( element ) => (
-				<Ariakit.CompositeHover
-					store={ compositeStore }
+				<Composite.Hover
 					key={ element.value }
 					render={
-						<CompositeItem
+						<Composite.Item
 							render={
 								<div
 									aria-label={ element.label }
@@ -192,7 +188,7 @@ function ListBox( { view, filter, onChangeView }: SearchWidgetProps ) {
 							</span>
 						) }
 					</span>
-				</Ariakit.CompositeHover>
+				</Composite.Hover>
 			) ) }
 		</Composite>
 	);